+++ sdk/communication/azure-communication-callautomation/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Mid Call actions support overriding callback url.
 - Cancel adding Participant invitation.
 - Support transfer a participant in a group call to another participant.
+- Add Custom Context payload to Transfer and AddParticipant API.
 
 ### Breaking Changes